requirements.txt
streamlit>=1.40.0
google-genai>=1.0.0
tavily-python>=0.3.3
python-dotenv>=1.0.0
